Lotto 자동 구매, 수동구매 service 수정
